Creating various files
I have created several files to work with and on it.
The folder "extra src" contains files that are not working or simply contains useful code to use.

"magnebot_ejemplos.py" is the table with the puzzle box
"prueba magnebot.py" doesn't work, but maybe useful
 "magnebot_paco.py" places a magnebot a turn it his torso
"magnebot2.py" places a magnebot and goes forward and then backward

"robot_camera.py" is a example to place a camera in the robot's torso and see through it